requirements.txt
langchain==0.1.20
langchain_chroma==0.1.1
langchain_community==0.0.38
langchain_core==0.2.5
langchain_openai==0.1.6
python-dotenv==1.0.1
PyPDF2==3.0.1
tqdm==4.66.4